refactor: rename ExoPlayerBitrateLogger to bitrateHandler
